app/requirements.txt
matplotlib
keras-cv
uuid7
boto3
websocket-client
loguru
rel
nats-py
diffusers
transformers